% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Layout.R
\name{AddLayout.Benchmark}
\alias{AddLayout.Benchmark}
\title{Add 2-dimensional layout to a Benchmark object}
\usage{
\method{AddLayout}{Benchmark}(
  benchmark,
  idx.subpipeline = NULL,
  idx.n_param = NULL,
  method = NULL,
  idcs_training = NULL,
  coordinates = NULL,
  ask_overwrite = TRUE,
  seed = 1,
  verbose = TRUE
)
}
\description{
Add 2-dimensional layout to a Benchmark object
}
