% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_BenchmarkSetUp_.R
\name{Benchmark}
\alias{Benchmark}
\title{\code{Benchmark} object constructor}
\usage{
Benchmark(
  input,
  input_labels = NULL,
  hierarchy = FALSE,
  hierarchy.params = list(g_c = 0, g_a = 0.2, m_c = 0.4, m_a = 0),
  input_features = NULL,
  remove_labels = NULL,
  unassigned_labels = c("unassigned"),
  compensation_matrix = NULL,
  transform = NULL,
  transform_cofactor = NULL,
  transform_features = NULL,
  input_marker_types = NULL,
  batch_id = "sample_id",
  projection.training_set = NULL,
  clustering.training_set = NULL,
  precomputed_knn = NULL,
  knn.k = 100,
  knn.distance = "euclidean",
  knn.features = NULL,
  precomputed_dist = NULL,
  pipeline = NULL,
  n_params = NULL,
  stability_repeat = NULL,
  stability_bootstrap = NULL,
  benchmark_name = "SingleBench_Benchmark",
  h5_path = paste0(benchmark_name, ".h5"),
  ask_overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{input}{string vector, \code{flowSet} or \code{SummarizedExperiments}: input dataset given as vector of FCS file paths, \code{flowSet} object or \code{SummarizedExperiment} object}

\item{input_labels}{optional string or factor vector (or a list of these vectors): if manual labels of data points are not included in the input FCS file(s), \code{flowSet} or \code{SummarizedExperiment}, specify them here. For a single FCS file or a \code{flowSet} or \code{SummarizedExperiment}, use a single vector (of length equal to number of expression matrix rows).
If multiple FCS files are used as input, use a list of vectors, one for each of the files (in the same order as the FCS file paths).}

\item{hierarchy}{logical: whether to use hierarchical penalties for mismatches in scoring clustering results. Default value is \code{FALSE}}

\item{hierarchy.params}{optional list: parameter values for hierarchical penalties model}

\item{input_features}{optional numeric or string vector: which columns of the input data should be used in analysis. Default value is \code{NULL}, which causes all columns to be used}

\item{remove_labels}{optional string vector: names of population labels to exclude from analysis. Default value is \code{NULL}}

\item{unassigned_labels}{optional string or string vector: names of population labels to handle as 'unassigned' events (not belonging to a manually annotated population) for purposes of computing evaluation metrics. Default value is '\code{unassigned}'}

\item{compensation_matrix}{optional numeric matrix: compensation matrix to use if input is uncompensated flow cytometry data. See \code{flowCore::compensate}. Default value is \code{NULL}, whereby compensation is not applied}

\item{transform}{optional string: name of transformation function to apply to input expression data. If specified, it must be one of \code{asinh}, \code{estimateLogicle}. Default value is \code{NULL}, whereby transformation is not applied}

\item{transform_cofactor}{optional numeric value: cofactor for a transformation function specified in \code{transform}. For example, \code{asinh} with cofactor \code{5} is usually appropriate for mass cytometry data. Default value is \code{NULL}}

\item{transform_features}{optional numeric or string vector: channels or indices of columns to transform (if \code{transform} is specified). Intersection between the indices given here and the indices retrieved from \code{idcs.input_features} or using \code{input_marker_types} are taken (if either of those two parameters is specified). Default value is \code{NULL}, which translates to 'all features that are used in the analysis'}

\item{input_marker_types}{optional string: if using \code{SummarizedExperiment} as input, which type(s) of markers (specified in vector \code{marker_type} of \code{colData}) should be used in analysis. This overrides \code{idcs.input_features}. Default value is \code{NULL}}

\item{batch_id}{optional string: if using \code{SummarizedExperiment} as input, which \code{rowData} column should be used to separate input data into batches if available. Default value is \code{sample_id}}

\item{projection.training_set}{optional vector of integers: if input dataset comprises multiple samples, which ones should be used as the training set for building a dimension-reduction model. Default value is \code{NULL} (DR is applied to the entire dataset)}

\item{clustering.training_set}{optional vector of integers: if input dataset comprises multiple samples, which ones should be used as the training set for building a clustering model. Default value is \code{NULL} (clustering is applied to the entire dataset)}

\item{precomputed_knn}{optional list: a previously computed \code{k}-NN matrix for input data as a list containing slots \code{Indices} and \code{Distances}.}

\item{knn.k}{integer: if data denoising or the evaluation of DR or clustering tools requires generation of a \code{k}-NN matrix, what value of \code{k} should be used. Default value is \code{100}}

\item{knn.distance}{string: if \code{k}-NN matrix needs to be constructed, which distance metric should be used. The only option currently is '\code{euclidean}'. Default value is '\code{euclidean}'}

\item{knn.features}{optional numeric or string vector: channels or indices of columns to use in construction of \code{k}-NN matrix. Default value is \code{NULL}, which causes the \code{k}-NN algorithm to use \code{idcs.transform_features}}

\item{precomputed_dist}{optional numeric matrix: a previously computed distance matrix. Default value is \code{NULL}}

\item{pipeline}{list of \code{Subpipeline}s}

\item{n_params}{list of named \emph{n}-parameter lists for each subpipeline (eg. \code{list(list(projection = c(20, 10), clustering = c(30, 30)))})}

\item{stability_repeat}{optional integer value between \code{2} and \code{1000}: number of repeated runs of each clustering step on full input. Default value is \code{NULL}}

\item{stability_bootstrap}{optional integer value between \code{2} and \code{1000}: number of repeated runs of each clustering step on different bootstraps of input, in addition to one final run on full data. Default value is \code{NULL}. Overrides \code{cluster.stability_randomstart}}

\item{benchmark_name}{string: benchmark name. Default value is '\code{SingleBench_Benchmark}'}

\item{h5_path}{string: name of auxiliary HDF5 file to generate. Default value is \code{benchmark_name} (with an '\code{.h5}' extension)}

\item{ask_overwrite}{logical: if \code{h5_path} is specified and an HDF5 file with the specified name exists already, should the user be asked before the file is overwritten? Default value is \code{TRUE}}

\item{verbose}{logical value: specifies if informative messages should be produced throughout construction of the \code{Benchmark} object. Default value is \code{TRUE}}
}
\description{
Creates an object of type \code{Benchmark}, used to configure a benchmarking pipeline and input data.
This object allows to run projection (dimension-reduction/denoising) and clustering pipelines on high-dimensional single-cell data (such as from \strong{flow cytometry}, \strong{mass cytometry}, \strong{CITE-seq}, \strong{scRNA-seq}).
}
\details{
Once set up, use \code{Evaluate} to run the benchmark.
}
\section{Input data and tools}{
Input to the pipeline (expression data matrix/matrices) is passed as a path to an FCS file (or multiple FCS files) or a \code{flowSet} or \code{SummarizedExperiment} object, using the \code{input} argument.
Projection and clustering tools (that should be applied to the data upon evaluation) are passed to the constructor using modules (see the \code{Module} constructor).
}

\section{Manual labelling of data points}{
Each input data point needs to have a label, assigning it to a population.
Labels can either be included in the input data object or supplied separately in a vector (\code{input_labels}).

Including labels directly is only possible with a \code{flowSet} or \code{SummarizedExperiment} (see convention in package \code{HDCytoData}).
With a \code{flowSet}, include a '\code{population_id}' channel with numeric flags and mapping of sorted numeric flags to population names, respectively, as columns of a data frame accessible for each \code{flowFrame} via \code{@description$POPULATION_INFO}.
With a \code{SummarizedExperiment}, include a vector of population labels directly as the first column of \code{rowData}.
In addition, \code{colData} should contain a \code{channel_name}, \code{marker_name} and (optionally) a \code{marker_class} slot.
}

\section{Subpipelines and \emph{n}-parameters}{
A pipeline in made up of subpipelines, which are combinations of tools and their parameters.
Each subpipeline can have one or both of two modules: projection and clustering.
For instance, a single subpipeline for data smoothing and subsequent \code{FlowSOM} clustering is created using:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ subpipelines <- list()
 subpipelines[[1]] <- 
     Subpipeline(
         projection = list(Module(Fix('smooth', k = 50), n_param = 'n_iter')),
         clustering = Module(Fix('FlowSOM', grid_width = 25, grid_height = 25), n_param = 'n_clusters')
     )
}\if{html}{\out{</div>}}

Up to two \emph{n}-parameters (one per projection step and one per clustering step) can be specified.
This means that, within the subpipeline, we can do parameter sweeps over multiple combinations of values of these parameters,
These values need to be specified in the \code{Benchmark} constructor.
For instance, to try out multiple latent-space dimensions and target cluster counts, we can specify:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  n_params <- list()
  n_params[[1]] <- list(
      projection = rep(c(0, 1, 2, 4), each = 2),
      clustering = c(30, 40)
 )
}\if{html}{\out{</div>}}

If you want to set up a second subpipeline that re-uses the projection step, you can simply clone the same set-up:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ subpipeline[[2]] <-
     Subpipeline(
         projection = CloneFrom(1),
         clustering = Module(Fix('Depeche'), n_param = 'fixed_penalty')
     )
 
 n_params[[2]] <- list(
     projection = rep(c(0, 1, 2, 4, 5), each = 3),
     clustering = c(2, 3, 4)
 )
}\if{html}{\out{</div>}}

Since we used \code{CloneFrom}, any results that have been produced during the first subpipeline and are also used in the second one will be recycled (they are not produced again nor are they actually copied).

Of course \emph{n}-parameter do not need to be specified, in which case they are simply not used (only one result is produced for the subpipeline).
Similarly, only one subpipeline per pipeline can be generated (leading to a simgle result of the pipeline).
}

\section{Stability analysis of clustering}{
If you are interested in analysing stability of a clustering algorithm, run the tool multiple times (with a different random seed) or apply it multiple times to a bootstrap of the input data.
That way, for your clustering evaluation metrics (scores) you will get a range of values (instead of a single values) for each subpipeline and \emph{n}-parameter iteration.
For repeated runs, use the \code{stability_repeat} parameter.
For bootstraps, use the \code{stability_bootstrap} parameter.
}

\section{Train-and-map}{
Some projection and clustering tools allow you to train a model on a subset of your input data and map the results onto the rest.
If your input dataset consists of multiple FCS files, \code{flowFrame}s or \code{SummarizedExperiment} batches, you can specify the training set indices in \code{projection.training_set} and \code{clustering.training_set}.
}

\section{Hierarchical labelling}{
If the manual labels of your input data are derived from a hierarchy of populations (ie. \strong{gating hierarchy} for cytometry data), you can make use of the entire hierarchy for evaluation purposes.
For instance, instead of using a '\strong{CD4+ T cell}' label, you can use '\strong{Lymphocyte/T cell/CD4+ T cell}' (using a path-like label with '\code{/}' as separator).
Then, if you apply a clustering tool and match each cluster to a population present in the data, \code{SingleBench} can evaluate the quality of clustering more carefully.
Specifically, instead of distinguishing between \emph{match} versus \emph{mismatch}, a scoring matrix is produced which penalises mismatches with different severity.
For instance, to misclassify '\strong{Lymphocyte/T cell/CD4+ T cell}' as '\strong{Lymphocyte/T cell}' can be better than misclassifying it as '\strong{Lymphocyte/T cell/CD8+ T cell}', which is still better than misclassifying it as '\strong{Lymphocyte/B cell/Alpha-Beta Mature B Cell}'.

The scoring of each potential mismatch is based on the route from the true label to the predicted label through the label hierarchy tree.
To parametrise the hierarchical penalty model, you can set 4 custom values.
Firstly, the '\emph{constant generalisation penalty}' \code{g_c} penalises the first step taken in the direction of the tree root and the '\emph{additive generalisation penalty}' \code{g_a} penalises each step in that direction.
Secondly, the '\emph{constant misidentification penalty}' \code{m_a} panelises the first step taken in the direction of the tree leaves and the '\emph{additive misidentification penalty}' \code{m_a} penalises each step in that direction.
The values of these penalties are positive values, and the sum of penalties for a misclassification get subtracted from \code{1}, which is the score for correct match.
By default, \code{g_c = 0}, \code{g_a = 0.2}, \code{m_c = 0.4}, \code{m_a = 0}.

If you want to use hierarchical penalties, make sure you use the full labels (paths using '\code{/}' as separator), set the argument \code{hierarchy} to \code{TRUE}.
Optionally, set \code{hierarchy.params} to a named list with slots \code{g_c}, \code{g_a}, \code{m_c} and \code{m_a}.
Alternatively, instead of computing a scoring matrix for the penalties automatically, you can include a slot \code{scoring_matrix}, which contains the names of all labels as column and row names and a match/mismatch score for each combination of true and predicted label.
This scoring matrix will then be used directly and will override any other paramters of the model.
}

\section{AWDQrTr5v5sNiYj1b2bjYxKZ5ZGhZ9ce-49--NN matrix construction}{
Various dimension-reduction or denoising algorithms may require the construction of a \emph{k}-nearest-neighbour graph.
If any of them are included in the pipeline, a single \code{k}-NN matrix will be constructed at the start of evaluation.
You can specify \code{k} (neighbour count) for k-NNG construction and the distance metric as \code{knn.k} and \code{knn.distance}
}

\section{Distance matrix construction}{
You can decide to score the results of projection steps using full distance matrices (distances of each point to each other point) of the original input matrix and each produced projection.
This can be set up in the \code{Evaluate} function.
(Beware that construction of a full distance matrix has quadratic complexity with regard to number of points.)
To speed things up, you can load a pre-computed distance matrix of your original input data by specifying the value of parameter \code{precomputed_dist}.
To compute such a distance matrix you can use \code{as.matrix(stats::dist(.))} or \code{coRanking:::euclidean_C(.)}.
(However, know that even if you do provide a pre-computed distance matrix of the original data, distance matrices for each projection produced in the benchmark pipeline will still need to be computed, and this can take a long time for larger datasets.)
}

\section{HDF5 integration}{
All large vectors and matrices generated during benchmark set-up and evaluation are stored in an auxiliary HDF5 file.
Erasing it amounts to discarding results of the benchmark.
}

\section{Creating new wrappers}{
To create new wrappers for projection and clustering tools, use \code{WrapTool}.
}

\seealso{
\itemize{
\item \strong{\code{Evaluate}}: runs all benchmark subpipelines and scores the performance of each tool
\item \strong{\code{AddLayout}}: allows you to add a separate 2-dimensional layout of the input dataset or to use an existing projection (produced in the evaluation) as a visualisation layout.
}
}
