% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_BenchmarkSetUp_00_Modules.R
\name{CloneFrom}
\alias{CloneFrom}
\title{Clone projection step from some other subpipeline}
\usage{
CloneFrom(idx_subpipeline)
}
\arguments{
\item{idx_subpipeline}{integer: index of subpipeline from which to clone the projection result}
}
\description{
If the set-up of a projection step is shared among multiple subpipelines, only specify it once and use this cloning function to refer to the first instance.
}
