% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Evaluate_kNN.R
\name{ComputekNNMatrix}
\alias{ComputekNNMatrix}
\title{Compute a \code{k}-nearest-neighbour matrix for expression data}
\usage{
ComputekNNMatrix(exprs, k, metric = "euclidean", out.systime = NULL)
}
\arguments{
\item{exprs}{numeric matrix: a coordinate matrix of biological expression data (columns correspond to markers, rows correspond to cells)}

\item{k}{integer: number of nearest neighbours to find for each point (not including self)}

\item{metric}{string: distance metric to pass to \code{RcppHNSW::hnsw_knn}. Default value is \code{'euclidean'}}

\item{out.systime}{optional out-variable: if an object is passed as \code{out.systime}, a side-effect of executing this function is that this object will be assigned elapsed time (in seconds) needed to complete the \code{k}-NN search}
}
\value{
list with two slots: \code{Indices} contains a matrix of nearest neighbours to each point (per row) and \code{Distances} contains a matrix of corresponding Euclidean distances
}
\description{
Finds \code{k} closest neighbours to each point in high-dimensional space and the distances to these points.
The approximate k-NN search algorith from \code{RcppHNSW} is used.
}
