% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Evaluate_.R
\name{Evaluate.Benchmark}
\alias{Evaluate.Benchmark}
\title{Evaluate a benchmark}
\usage{
\method{Evaluate}{Benchmark}(
  benchmark,
  score_projections = FALSE,
  projection_neighbourhood = 100,
  n_cores = NULL,
  which_python = NULL,
  seed.projection = 1,
  seed.clustering = 1,
  ask_overwrite = TRUE,
  verbose = TRUE
)
}
\description{
Evaluate a benchmark
}
