% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_kNNGating.R
\name{ExtendGates}
\alias{ExtendGates}
\title{Extend gating to unassigned cells}
\usage{
ExtendGates(
  exprs,
  annotation,
  unassigned_labels,
  knn_indices,
  k = NULL,
  n_iter = 1,
  verbose = TRUE
)
}
\arguments{
\item{exprs}{numeric matrix: a coordinate matrix of biological expression data (columns correspond to markers, rows correspond to cells)}

\item{annotation}{string or factor vector: vector of labels per each row of \code{exprs}}

\item{unassigned_labels}{string or string vector: names of population labels to handle as 'unassigned' events (not belonging to a manually annotated population)}

\item{knn_indices}{numeric matrix: the \code{Indices} slot of a \code{k}-NN object produced by the function \code{ComputekNNMatrix}}

\item{k}{integer: number of nearest neighbours of each point to use for labelling. Default value is the maximum possible value (\code{k} of the \code{k}-NN matrix)}

\item{n_iter}{integer: number of labelling iterations. Default value is \code{1}}

\item{verbose}{logical: whether to display progress messages. Default value is \code{TRUE}}
}
\description{
Uses a \code{k}-NN classifier to label cells that are marked as 'unassigned' in an existing manual annotation vector.
A majority vote is taken among nearest neighbours to an unassigned cells, excluding other unassigneds.
This procedure can be repeated in multiple consecutive iterations.
}
\seealso{
\itemize{
\item \strong{\code{ComputekNNMatrix}}: finds \code{k} nearest neighbours to each point in high-dimensional expression data
}
}
