% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_BenchmarkSetUp_00_Modules.R
\name{Fix}
\alias{Fix}
\title{Fix parameters for a tool wrapper}
\usage{
Fix(tool_name, ...)
}
\arguments{
\item{tool_name}{name of a tool for which a wrapper (generated by \code{WrapTool}) exists in the global namespace}

\item{...}{named parameter values to fix for training a model using \code{wrapper}}
}
\description{
Takes a projection or clustering tool wrapper and fixes given parameter values for it.
}
