% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetAnnotation}
\alias{GetAnnotation}
\title{Get \code{Benchmark} manual annotation}
\usage{
GetAnnotation(benchmark, concatenate = FALSE)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{concatenate}{logical value; if a list of annotation vectors (per sample) is available, should it be concatenated into one vector? Default value is \code{FALSE}}
}
\description{
Extracts manual annotation vector (or list of vectors) associated with inputs to a benchmark pipeline.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
