% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetAnnotationScoring}
\alias{GetAnnotationScoring}
\title{Get \code{Benchmark} manual annotation scoring}
\usage{
GetAnnotationScoring(benchmark, suffix = "")
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{suffix}{optional string (advanced): HDF5 slot name suffix}
}
\description{
Extracts values of unsupervised evaluation metrics applied to manual annotation of benchmark input data.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
