% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetBootstrapIndices}
\alias{GetBootstrapIndices}
\title{Get \code{Benchmark} bootstrap indices}
\usage{
GetBootstrapIndices(benchmark)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}
}
\description{
Extracts vector of indices (or list of vectors of indices) used for taking bootstraps of input expression data for clustering (if this is done).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)\
}
}
