% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AlignChannels.R
\name{GetChannelsAndMarkers}
\alias{GetChannelsAndMarkers}
\title{Extract channel and marker names from a collection of FCS files}
\usage{
GetChannelsAndMarkers(fcs_paths, verbose = TRUE)
}
\arguments{
\item{fcs_paths}{string vector: paths to FCS files}

\item{verbose}{logical value: whether progress messages should be printed during execution}
}
\value{
list with slots \code{markers}, \code{channels}, \code{pairs} (string representation of how markers and channels are matched) and \code{idcs.mismatches} (which parameters are mismatched across the different files).
}
\description{
For a collection of FCS files from one experiment (same panel), this function extracts names of channels and markers associated with them, as well as identifying possible mismatches or unmatched parameters that are not present in all files.
}
