% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetClustering}
\alias{GetClustering}
\title{Get \code{Benchmark} clustering result}
\usage{
GetClustering(
  benchmark,
  idx.subpipeline,
  idx.n_param = NULL,
  idx.run = NULL,
  concatenate = FALSE
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark}}

\item{idx.n_param}{integer: index of an \emph{n}-parameter iteration of a sub-pipeline of \code{benchmark}}

\item{idx.run}{integer or integer vector: run of clustering algorithm to extract results from. Default value is \code{NULL}, which extracts a list of results from all runs}
}
\description{
Extracts the result of a clustering tool that is part of a benchmark sub-pipeline that was set up previously.
If stability analysis is turned on for the benchmark, results from one or more repeated runs can be returned.
}
\details{
If you choose to extract results of all runs of a clustering algorithm which clustered on bootstraps of the original data, the single run on original data is returned.
Furthermore, you can extract bootstrap indices using \code{GetBootstrapIndices}.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)\
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
