% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetClusteringInput}
\alias{GetClusteringInput}
\title{Get \code{Benchmark} clustering input}
\usage{
GetClusteringInput(
  benchmark,
  idx.subpipeline,
  idx.n_param = NULL,
  null_if_exprs = FALSE
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark}}

\item{idx.n_param}{integer: index of an \emph{n}-parameter iteration of a sub-pipeline of \code{benchmark}}

\item{null_if_exprs}{logical: whether to return \code{NULL} instead of original expression data. Default value is \code{FALSE}}
}
\description{
Extracts the input to a clustering tool that is part of a benchmark sub-pipeline that was set up previously.
If a dimension-reduction tool precedes the clustering tool, the latent-space projection from that tool is retrieved.
Otherwise, the original expression data is retrieved.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)\
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
