% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetClusteringScoringTable}
\alias{GetClusteringScoringTable}
\title{Get \code{Benchmark} clustering evaluation scores table}
\usage{
GetClusteringScoringTable(benchmark, idx.subpipeline)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark} that includes a clustering step}
}
\description{
Returns a \code{tibble} of clustering evaluation metric values for specified subpipeline of a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
