% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetCoRanking}
\alias{GetCoRanking}
\title{Get \code{Benchmark} collapsed co-ranking matrix}
\usage{
GetCoRanking(benchmark, idx.subpipeline, idx.n_param = NULL)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer value: index of subpipeline that includes a projection step}

\item{idx.n_param}{optional integer value: index of subpipeline n-parameter iteration. Default value is \code{NULL}}
}
\description{
Extracts the collapsed co-ranking matrix for assessing quality of a projection.
}
