% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetDistanceMatrix}
\alias{GetDistanceMatrix}
\title{Get \code{Benchmark} expression data distance matrix}
\usage{
GetDistanceMatrix(benchmark)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}
}
\description{
Extracts distance matrix associated with inputs to a benchmark pipeline.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
