% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetExpressionMatrix}
\alias{GetExpressionMatrix}
\title{Get \code{Benchmark} expression data}
\usage{
GetExpressionMatrix(benchmark, concatenate = FALSE)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{concatenate}{logical value; if a list of layout matrices (per sample) is available, should it be concatenated into one vector? Default value is \code{FALSE}}
}
\description{
Extracts expression matrix (or list of expression matrices) associated with inputs to a benchmark pipeline.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
