% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetkNNMatrix}
\alias{GetkNNMatrix}
\title{Get \code{Benchmark} expression data \code{k}-NN matrix}
\usage{
GetkNNMatrix(benchmark)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}
}
\description{
Extracts \emph{k}-nearest-neighbours matrix associated with inputs to a benchmark pipeline.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
