% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetLayout}
\alias{GetLayout}
\title{Get \code{Benchmark} 2-dimensional layout of input data}
\usage{
GetLayout(benchmark, concatenate = FALSE)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{concatenate}{logical value; if a list of layout matrices (per sample) is available, should it be concatenated into one vector? Default value is \code{FALSE}}
}
\description{
Extracts coordinate matrix of a 2-dimensional layout associated with the input data for a \code{Benchmark} object (benchmark pipeline set-up).
This layout can be created by applying \code{AddLayout} to a previously evaluated benchmark pipeline and serves visualisation purposes.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)\
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
\item \strong{\code{AddLayout}}: allows you to add a separate 2-dimensional layout of the input dataset or to use an existing projection (produced in the evaluation) as a visualisation layout.
}
}
