% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetNParameterIterationName}
\alias{GetNParameterIterationName}
\title{Get \code{Benchmark} \emph{n}-parameter iteration name}
\usage{
GetNParameterIterationName(
  benchmark,
  idx.subpipeline,
  idx.n_param,
  with_tool_names = TRUE
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a subpipeline of \code{benchmark}}

\item{idx.n_param}{integer: index of an \emph{n}-parameter iteration of a sub-pipeline of \code{benchmark}}

\item{with_tool_names}{logical: whether tool name(s), in addition to \emph{n}-parameter value(s), should be included in the name. Default value is \code{TRUE}}
}
\description{
Extracts name of an \emph{n}-parameter iteration of a subpipeline in a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
