% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetNParameterIterationName_Clustering}
\alias{GetNParameterIterationName_Clustering}
\title{Get \code{Benchmark} \emph{n}-parameter iteration clustering name}
\usage{
GetNParameterIterationName_Clustering(
  benchmark,
  idx.subpipeline,
  idx.n_param,
  with_tool_name = TRUE
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a subpipeline of \code{benchmark}}

\item{idx.n_param}{integer: index of an \emph{n}-parameter iteration of a sub-pipeline of \code{benchmark}}

\item{with_tool_name}{logical: whether tool name, in addition to \emph{n}-parameter value, should be included in the name. Default value is \code{TRUE}}
}
\description{
Extracts name of an \emph{n}-parameter iteration of the clustering part of a subpipeline in a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
