% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetNParameterIterationsCount}
\alias{GetNParameterIterationsCount}
\title{Get \code{Benchmark} \emph{n}-parameter count for subpipeline}
\usage{
GetNParameterIterationsCount(benchmark, idx.subpipeline)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark}}
}
\description{
Extract the number of \emph{n}-parameter iterations for a chosen subpipeline in a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
