% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetNParameterNames}
\alias{GetNParameterNames}
\title{Get \code{Benchmark} \emph{n}-parameter iteration names of a subpipeline}
\usage{
GetNParameterNames(benchmark, idx.subpipeline, with_tool_names = FALSE)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a subpipeline of \code{benchmark}}

\item{with_tool_names}{logical: whether names of the modules should be included. Default value is \code{FALSE}}
}
\description{
Extracts names of all \emph{n}-parameter iterations of a subpipeline in a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
