% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetNParameterValues}
\alias{GetNParameterValues}
\title{Get \code{Benchmark} \emph{n}-parameter values for a subpipeline}
\usage{
GetNParameterValues(benchmark, idx.subpipeline)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a subpipeline of \code{benchmark}}
}
\description{
Extracts combinations of \emph{n}-parameter values for each \emph{n}-parameter iteration of a subpipeline in a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
