% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetPenaltyScoringMatrix}
\alias{GetPenaltyScoringMatrix}
\title{Get \code{Benchmark} penalty scoring matrix}
\usage{
GetPenaltyScoringMatrix(benchmark)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark} that uses hierarchical penalties}
}
\description{
Extracts the penalty scoring matrix, which defines a hierarchy penalty model, from a \code{Benchmark}-type object.
}
\seealso{
\itemize{
\item \strong{\code{CreatePenaltyScoringMatrix}}: creates a penalty scoring matrix for a given set of manually annotated populations
}
}
