% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetProjection}
\alias{GetProjection}
\title{Get \code{Benchmark} latent-space projection}
\usage{
GetProjection(benchmark, idx.subpipeline, idx.n_param = NULL)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark}}

\item{idx.n_param}{integer: index of an \emph{n}-parameter iteration of a sub-pipeline of \code{benchmark}}
}
\description{
Extracts coordinate matrix of a latent-space projection generated by evaluating an object of type \code{Benchmark}.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)\
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
