% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetProjectionScores}
\alias{GetProjectionScores}
\title{Get \code{Benchmark} dimension-reduction evaluation scores}
\usage{
GetProjectionScores(benchmark, idx.subpipeline, idcs.n_param = NULL)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a subpipeline of \code{benchmark}}

\item{idx.n_param}{integer: index of an \emph{n}-parameter iteration of a sub-pipeline of \code{benchmark}}
}
\description{
Extracts list of evaluation scores and other products of evaluating performance of a projection step of a sub-pipeline \emph{n}-iteration that includes dimension reduction.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
