% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetProjectionScoringTable}
\alias{GetProjectionScoringTable}
\title{Get \code{Benchmark} projection evaluation scores table}
\usage{
GetProjectionScoringTable(benchmark, idx.subpipeline)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a subpipeline of \code{benchmark} that includes a projection step}
}
\description{
Returns a \code{tibble} of projection evaluation metric values for specified subpipeline of a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
\item \strong{\code{Evaluate}}: runs all benchmark subpipelines and scores the performance of each tool
}
}
