% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetRMSDPerCluster}
\alias{GetRMSDPerCluster}
\title{Get RMSDs of each cluster in clustering result of an evaluated \code{Benchmark} pipeline}
\usage{
GetRMSDPerCluster(
  benchmark,
  idx.subpipeline,
  idx.n_param = NULL,
  idx.run = NULL
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark} that includes a clustering step}

\item{idx.n_param}{integer: index of \emph{n}-parameter iteration of a subpipeline of \code{benchmark} (if \emph{n}-parameters specified). Default value is \code{NULL}}

\item{idx.run}{integer: index of clustering run if repeated runs were evaluated. Default value is 1}
}
\description{
Returns the vector of RMSDs per cluster.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
