% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetScores}
\alias{GetScores}
\title{Get \code{Benchmark} evaluation scores of an \emph{n}-parameter iteration}
\usage{
GetScores(benchmark, idx.subpipeline, idcs.n_param = NULL)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark}}

\item{idcs.n_param}{integer or vector of integers: indices of \emph{n}-parameter iterations of a subpipeline of \code{benchmark}}
}
\description{
Extracts list of evaluation scores and other products of evaluating performance of tools used in a subpipeline \emph{n}-iteration in a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
