% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetSubpipelineName}
\alias{GetSubpipelineName}
\title{Get \code{Benchmark} sub-pipeline name}
\usage{
GetSubpipelineName(benchmark, idx.subpipeline)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark}}
}
\description{
Extracts name of a sub-pipeline in a \code{Benchmark} object (benchmark pipeline set-up).
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)
}
}
