% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_BenchmarkSetUp_00_Modules.R
\name{Module}
\alias{Module}
\title{Specify \emph{n}-parameter for a tool wrapper}
\usage{
Module(wrapper_with_parameters, n_param = NULL)
}
\arguments{
\item{n_param}{optional string: name of the \emph{n}-parameter}

\item{wrapper_with_parameter}{tool wrapper with parameters generated by \code{Fix}}
}
\description{
Takes a projection or clustering tool wrapper with parameter values and chooses which (if any) of the parameters to turn into an \emph{n}-parameter (non-fixed, to do a parameter sweep over).
}
