% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_Heatmaps.R
\name{PlotClusterHeatmap}
\alias{PlotClusterHeatmap}
\title{Create a heatmap of expression value medians per cluster}
\usage{
PlotClusterHeatmap(
  benchmark,
  idx.subpipeline,
  idx.n_param = NULL,
  idx.run = 1,
  scale = "none",
  display_numbers = FALSE,
  ...
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate}}

\item{idx.subpipeline}{integer value: index of subpipeline that includes a clustering step}

\item{idx.n_param}{integer: index of \emph{n}-parameter(s) iteration of interest. Default value is \code{NULL}}

\item{idx.run}{integer: index of run if repeated. Default value is \code{1}}

\item{scale}{string: method of scaling heatmap colours ('row', 'column' or 'none). Default value is 'none'}

\item{display_numbers}{logical: whether to display numbers in the heatmap. Default value is \code{FALSE}}

\item{...}{additional arguments to be passed to \code{pheatmap::pheatmap}}
}
\description{
Creates a heatmap showing expression value medians per each cluster, taking clusters from a chosen subpipeline (and \emph{n}-parameter iteration(s)).
If multiple \emph{n}-parameter iterations are chosen, a list of plots is generated.
}
