% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_Overlays.R
\name{PlotClustersOverlay}
\alias{PlotClustersOverlay}
\title{Create a clusters-overlay plot}
\usage{
PlotClustersOverlay(
  benchmark,
  idx.subpipeline,
  idx.n_param = NULL,
  idx.run = 1,
  exclude_unassigned = FALSE,
  raster_threshold = 5000L,
  plot_title = "Clustering",
  cluster_labels = FALSE,
  labelsize = 3,
  palette = c(RColorBrewer::brewer.pal(8, "Dark2"), RColorBrewer::brewer.pal(12,
    "Paired"), RColorBrewer::brewer.pal(12, "Set3")),
  svg_pointsize = 0.5,
  raster_pointsize = 0.005,
  fix_aspect_ratio = FALSE
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of sub-pipeline that includes a clustering step}

\item{idx.n_param}{integer: index of \emph{n}-parameter of interest. Default value is \code{NULL}}

\item{idx.run}{integer: index of repeated run of interest. Default value is 1}

\item{exclude_unassigned}{logical: if \code{TRUE}, data points that are considered unassigned per manual annotation are omitted. Default value is \code{FALSE}}

\item{raster_threshold}{integer: maximum number of data points for which vector graphics should be used. Default value is \code{5000}}

\item{plot_title}{string: title of the plot. Default value is '\emph{Clustering}'}

\item{cluster_labels}{logical: whether to label clusters within the plot. Default value is \code{FALSE}}

\item{labelsize}{numeric: size of the cluster labels within the plot, if shown. Default value is \code{3}}

\item{palette}{string vector: palette of distinctive colours for each cluster. Defaults to a palette of 32 colours}

\item{svg_pointsize}{numeric: size of the data points in the plot when using vector graphics. Default value is \code{0.5}}

\item{raster_pointsize}{numeric: size of the data points in the plot when using raster graphics. Default value is \code{0.005}}

\item{fix_aspect_ratio}{logical: whether to maintain a 1:1 aspect ratio of the plot. Default value is \code{FALSE}}
}
\description{
Creates a plot showing separation of generated clusters using previously generated 2-dimensional layout of input expression data.
For large numbers of data points, this function uses raster graphics to visualise the layout faster.
}
\details{
You need to specify a single sub-pipeline and one or more \emph{n}-parameter iteration by index.
If multiple \emph{n}-parameter iterations are chosen, a list of plots will be generated.

You can choose a custom title for your plot.
The plot subtitle is always the \emph{n}-parameter iteration name.
}
