% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_CoRanking.R
\name{PlotCoRanking}
\alias{PlotCoRanking}
\title{Plot collapsed co-ranking matrix for heatmap for projection in a benchmark}
\usage{
PlotCoRanking(benchmark, idx.subpipeline, idx.n_param = NULL, log = TRUE)
}
\arguments{
\item{benchmark}{an object of class \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate.Benchmark}}

\item{idx.subpipeline}{integer value: index of subpipeline that includes a projection step}

\item{idx.n_param}{optional integer value: index of subpipeline n-parameter iteration. Default value is \code{NULL}}
}
\description{
Creates plot showing collapsed co-ranking matrix of a dimension-reduction layout as a heatmap.
}
