% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_JaccardHeatmap.R
\name{PlotJaccardHeatmap}
\alias{PlotJaccardHeatmap}
\title{Create a comprehensive Jaccard heatmap plot}
\usage{
PlotJaccardHeatmap(benchmark, idx.subpipeline, idcs.n_param = NULL)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate}}

\item{idx.subpipeline}{integer value: index of sub-pipeline that includes a clustering step}

\item{idcs.n_param}{integer or vector of integers: indices of \emph{n}-parameter(s) (cluster count) of interest. Default value is \code{NULL}}
}
\value{
list of plots, one per each \emph{n}-parameter (cluster-count) iteration
}
\description{
Creates a comprehensive plot showing how the results of clustering map onto manually labelled populations.
Using this, you can look at which cell populations were identified correctly or incorrectly by automated clustering and what kind of mistakes the clustering tool made.
}
\details{
Moreover, the Jaccard heatmap shows eventual differences between matching clusters to populations bijectively (one-to-one) and taking the best cluster for each population (and vice versa).
A comparison of precision, recall and F1 scores for each match is also provided.

To create the plot, you need to specify which sub-pipeline and \emph{n}-parameter iteration(s) you want to look at.
If you choose multiple \emph{n}-parameter iterations, a list of plots is returned.
}
