% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_NParamMap.R
\name{PlotNParameterMap_Clustering}
\alias{PlotNParameterMap_Clustering}
\title{Plot an \emph{n}-parameter map for clustering}
\usage{
PlotNParameterMap_Clustering(benchmark, idx.subpipeline, nrow = NULL, ncol = 5)
}
\arguments{
\item{benchmark}{an object of class \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate.Benchmark}}

\item{idx.subpipeline}{integer value: index of subpipeline that includes a clustering step}

\item{nrow}{integer: optional specification of number of rows in the full plot. Default value is \code{NULL}}

\item{ncol}{integer: optional specification of number of columns in the full plot. Default value is \code{5}}
}
\description{
Creates plots showing values of clustering evaluation metrics and how they change with different DR or clustering \emph{n}-parameters.
If you choose a sub-pipeline where one \emph{n}-parameter (for instance, latent-space projection dimensionality or target cluster count) varies, a line plot is produced for each metrics.
If you choose one where both \emph{n}-parameters (for the projection step and the clustering step) vary, a heatmap is produced instead.
If neither \emph{n}-parameter varies, this function returns \code{NULL}.
}
