% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_Heatmaps.R
\name{PlotPopulationHeatmap}
\alias{PlotPopulationHeatmap}
\title{Create a heatmap of expression value medians per population}
\usage{
PlotPopulationHeatmap(benchmark, scale = "none", display_numbers = FALSE, ...)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}, as generated by the constructor \code{Benchmark}}

\item{scale}{string: method of scaling heatmap colours ('row', 'column' or 'none). Default value is 'none'}

\item{display_numbers}{logical: whether to display numbers in the heatmap. Default value is \code{FALSE}}

\item{...}{additional arguments to be passed to \code{pheatmap::pheatmap}}
}
\description{
Creates a heatmap showing expression value medians per each marker per each labelled population.
}
