% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_RMSDToSize.R
\name{PlotRMSDToSize}
\alias{PlotRMSDToSize}
\title{Plot RMSD-to-size plots for clustering result}
\usage{
PlotRMSDToSize(benchmark, idx.subpipeline, idx.n_param = NULL, idx.run = 1)
}
\arguments{
\item{benchmark}{an object of class \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate.Benchmark}}

\item{idx.subpipeline}{integer value: index of sub-pipeline that includes a projection step}

\item{idx.n_param}{integer value: index of \emph{n}-parameter iteration if \emph{n}-parameters were used. Deault value is \code{NULL}}

\item{idx.run}{integer value: number of run to select if multiple (repeated) runs of clustering algorithm are available. Default value is \code{1}}
}
\description{
Plots a grid of RMSD-to-size plots per manually assigned population.
This is an experimental tool to assess dissimilarities between manual labelling of data and the result of automated clustering.
}
\details{
RMSD (root mean square deviation) is a measure of the average variation in signal across all parameters.
A high RMSD will be seen in groups of points that span a large part of the high-dimensional space or are irregular in shape.
In the plotted grid, vertical axis shows RMSD and horizontal axis shows size.

The RMSD-to-size ratio is computed for each manually labelled population, each cluster mapped to it (by the fixed-cluster scheme) and the union of all the clusters matched to it.
Properties of clusters can be investigated based on their position in the grid relative to the matched (reference) population.
We assume that
\itemize{
\item clusters in the bottom-left quadrant (lower RMSD and smaller in size than the reference population) correspond to real sub-populations that go beyond the resolution of the manual annotation and are valuable to explore,
\item clusters in the bottom-right quadrant (lower RMSD and bigger in size) correspond to clusters that identified some different population, which nonetheless overlaps with the identified population,
\item clusters to the right of the reference population (with about the same RMSD and bigger in size) correspond to clusters that contain the matched population, but also include some unlabelled or mislabelled cells from it that might need to be included also,
\item clusters in the top-right quadrant (with higher RMSD and bigger in size) are a result from under-clustering,
\item clusters in the top-left quadrant (with higher RMSD and smaller in size) are a result of over-clustering.
}

Then for example, if a population mostly has clusters in the bottom-left quadrant matched to it, it seems the clustering identified smaller constituent populations that go beyond the resolution of manual labelling.
If a gating strategy has many unlabelled cells and we see clusters that are to the right of the reference population on the grid, these might likely correspond to unlabelled cells that really belong to the reference.
}
