% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_JaccardHeatmap.R
\name{PlotSimilarityHeatmap}
\alias{PlotSimilarityHeatmap}
\title{Create a comprehensive similarity heatmap plot}
\usage{
PlotSimilarityHeatmap(
  benchmark,
  idx.subpipeline = 1,
  idx.n_param = NULL,
  idx.run = 1,
  pheatmap_args = list()
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate}}

\item{idx.subpipeline}{integer value: index of sub-pipeline that includes a clustering step}

\item{idx.n_param}{integer: index of \emph{n}-parameter iteration of interest}

\item{idx.run}{integer: if clustering was run repeatedly for stability analysis, which run should be used to plot the heatmap}

\item{pheatmap_args}{named list: optional additional arguments to \code{pheatmap::pheatmap}}
}
\description{
Creates a comprehensive plot showing how the results of clustering map onto manually labelled populations using different matching strategies.
}
\details{
Using this, you can look at which cell populations were identified correctly or incorrectly by automated clustering and what kinds of mistakes the clustering set-up made.
Moreover, the heatmap shows eventual differences between matching clusters to populations bijectively (one-to-one) and taking the best cluster for each population (and vice versa).
A comparison of precision, recall and F1 scores for each match is also provided.

To create the plot, you need to specify which sub-pipeline and \emph{n}-parameter iteration you want to look at.
If you choose multiple \emph{n}-parameter iterations, a list of plots is returned.
}
