% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Scoring_Clustering_.R
\name{ScoreClustering}
\alias{ScoreClustering}
\title{Normalised Mutual Information'                          = nmi
Normalised Mutual Information'                                     = nmi
Normalised Mutual Information'                                        = nmi}
\usage{
ScoreClustering(
  exprs,
  annotation,
  result,
  stability,
  bootstrap_idcs,
  unassigned_labels,
  column_names
)
}
\description{
Normalised Mutual Information'                          = nmi
Normalised Mutual Information'                                     = nmi
Normalised Mutual Information'                                        = nmi
}
