% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_Readouts.R
\name{ShowReadout_Clustering}
\alias{ShowReadout_Clustering}
\title{View readout of clustering evaluation metric values}
\usage{
ShowReadout_Clustering(benchmark, idx.subpipeline)
}
\arguments{
\item{benchmark}{an object of class \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate.Benchmark}}

\item{idx.subpipeline}{integer value: index of sub-pipeline that includes a clustering step}
}
\description{
Shows a formatted table with values of clustering evaluation metrics for a chosen subpipeline of a \code{Benchmark} object.
If you are running R in \emph{RStudio}, the output should show up in the \emph{Viewer} pane.
Otherwise, you will probably see the ouput pop up in a new window/tab of your web browser.
}
