% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_BenchmarkSetUp_00_Modules.R
\name{Subpipeline}
\alias{Subpipeline}
\title{Create subpipeline of a pipeline}
\usage{
Subpipeline(projection = NULL, clustering = NULL)
}
\arguments{
\item{projection}{\code{Module} object for projection (can be composed of a single or multiple tools chained consecutively). Default value is \code{NULL}}

\item{clustering}{\code{Module} object for clustering (composed of a single clustering tool)}
}
\description{
Creates a single subpipeline, consisting of a projection step, a clustering step or both.
}
\seealso{
\code{Module}
}
