% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_WhatIs.R
\name{WhatIs}
\alias{WhatIs}
\title{Inspect a specific cluster or annotated population}
\usage{
WhatIs(
  benchmark,
  idx.subpipeline = 1,
  idx.n_param = NULL,
  idx.run = 1,
  cluster = NULL,
  population = NULL,
  marker1 = NULL,
  marker2 = NULL,
  pointsize_bg = 0.7,
  pointsize_fg = 1.2,
  exprs = NULL
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of subpipeline (if a cluster is to be retrieved). Default value is \code{NULL}}

\item{idx.n_param}{integer: \emph{n}-parameter iteration index (if a cluster is to be retrieved and \emph{n}-parameter was specified in the benchmark set-up). Default value is \code{NULL}}

\item{idx.run}{integer: which clustering run to use if clustering was run repeatedly. Default value is \code{1}}

\item{cluster}{integer: index of cluster of interest (if a cluster is to be retrieved). Default value is \code{NULL}}

\item{population}{string: name of population of interest (if a population is to be retrieved). Default value is \code{NULL}}

\item{marker1}{string: optional name of marker-of-interest on x-axis for biaxial expression plot. Default value is \code{NULL}}

\item{marker2}{string: optional name of marker-of-interest on y-axis for biaxial expression plot. Default value is \code{NULL}}

\item{pointsize_bg}{numeric: point size for plotting signal distribution from background (all) cells. Default value is \code{0.7}}

\item{pointsize_fg}{numeric: point size for plotting signal distribution from foreground (cluster-/population-specific) cells. Default value is \code{1.2}}
}
\description{
The identity of a cell subset (cluster or annotated population) is inspected in terms of its overall distribution of signal per marker, and correlations between markers at a single-cell level (indicating) cell-level patterns.
}
\details{
Alternatively, a biaxial expression plot for the given cell subset can be generated by specifying two markers of interest to plot against each other.
}
