% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Evaluate_kNN.R
\name{ComputekNNMatrix}
\alias{ComputekNNMatrix}
\title{Compute a \code{k}-nearest-neighbour matrix for expression data}
\usage{
ComputekNNMatrix(
  exprs,
  k,
  method = "annoy",
  dist = "euclidean",
  out.systime = NULL
)
}
\arguments{
\item{exprs}{numeric matrix: a coordinate matrix of biological expression data (columns correspond to markers, rows correspond to cells)}

\item{k}{integer: number of nearest neighbours to find for each point}

\item{method}{string: \code{k}-NN algorithm to use. Default value is \code{annoy}}

\item{out.systime}{optional out-variable: if an object is passed as \code{out.systime}, a side-effect of executing this function is that this object will be assigned elapsed time (in seconds) needed to complete the \code{k}-NN search}
}
\value{
list with two slots: \code{Indices} contains a matrix of nearest neighbours to each point (per row) and \code{Distances} contains a matrix of corresponding Euclidean distances
}
\description{
Finds \code{k} closest neighbours to each point in high-dimensional space and the distances to these points.
You can choose from various exact or approximate algorithms to use: \code{annoy}, \code{kd_tree}, \code{cover_tree}, \code{CR} and \code{brute}.
}
\references{
The code for \code{annoy} is adopted from \code{ivis}: \emph{Szubert, B., Cole, J.E., Monaco, C. et al. Structure-preserving visualisation of high dimensional single-cell datasets. Sci Rep 9, 8914 (2019). https://doi.org/10.1038/s41598-019-45301-0}
}
\seealso{
\itemize{
\item \strong{\code{Denoise}}: denoises high-dimensional expression data to drive down unwanted technical variation
}
}
