% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Getters.R
\name{GetClustering}
\alias{GetClustering}
\title{Get \code{Benchmark} clustering result}
\usage{
GetClustering(
  benchmark,
  idx.subpipeline,
  idx.n_param = NULL,
  all_runs = FALSE,
  concatenate = TRUE
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of a sub-pipeline of \code{benchmark}}

\item{idx.n_param}{integer: index of an \emph{n}-parameter iteration of a sub-pipeline of \code{benchmark}}

\item{all_runs}{logical: whether results of all runs should be returned (in a list). Default value is \code{FALSE}}
}
\description{
Extracts the result of a clustering tool that is part of a benchmark sub-pipeline that was set up previously.
If stability analysis is turned on for the benchmark, either results of all runs of the clustering can be returned, or only the result of a single run on the full expression data can be returned.
}
\details{
If you choose to extract results of all runs of a clustering algorithm which clustered on bootstraps of the original data, the last slot in the list contains the single run on original data.
Furthermore, you can extract bootstrap indices using \code{GetBootstrapIndices}.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)\
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
}
}
