% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_Heatmaps.R
\name{PlotClusterHeatmap}
\alias{PlotClusterHeatmap}
\title{Create a heatmap of expression value medians per cluster}
\usage{
PlotClusterHeatmap(benchmark, idx.subpipeline, idcs.n_param = NULL)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate}}

\item{idx.subpipeline}{integer value: index of subpipeline that includes a clustering step}

\item{idcs.n_param}{integer or vector of integers: indices of \emph{n}-parameter(s) iterations of interest. Default value is \code{NULL}, which translates to all \emph{n}-parameter iterations of the given sub-pipeline (or none if \emph{n}-parameters were not used)}
}
\description{
Creates a heatmap showing expression value medians per each cluster, taking clusters from a chosen subpipeline (and \emph{n}-parameter iteration(s)).
If multiple \emph{n}-parameter iterations are chosen, a list of plots is generated.
}
