% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_Overlays.R
\name{PlotClustersOverlay}
\alias{PlotClustersOverlay}
\title{Create a clusters-overlay plot}
\usage{
PlotClustersOverlay(
  benchmark,
  idx.subpipeline,
  idcs.n_param = NULL,
  exclude_unassigned = FALSE,
  raster_threshold = 5000L,
  plot_title = "Clustering"
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of sub-pipeline that includes a clustering step}

\item{idcs.n_param}{integer or vector of integers: indices of \emph{n}-parameter (cluster count) of interest. Default value is \code{NULL}, which translates to all \emph{n}-parameter iterations of the given sub-pipeline}

\item{exclude_unassigned}{logical: if \code{TRUE}, data points that are considered unassigned per manual annotation are omitted. Default value is \code{FALSE}}

\item{raster_threshold}{integer: maximum number of data points for which vector graphics should be used. Default value is \code{5000}}

\item{plot_title}{string: title of the plot. Default value is '\emph{Clustering}'}
}
\description{
Creates a plot showing separation of automatically generated clusters using previously generated 2-dimensional layout of input expression data.
For large numbers of data points, this function uses raster graphics to visualise the layout faster.
}
\details{
You need to specify a single sub-pipeline and one or more \emph{n}-parameter iteration by index.
If multiple \emph{n}-parameter iterations are chosen, a list of plots will be generated.

You can choose a custom title for your plot.
The plot subtitle is always the \emph{n}-parameter iteration name.
}
\seealso{
\itemize{
\item \strong{\code{Benchmark}}: constructs a \code{Benchmark} object (benchmark pipeline set-up)\
\item \strong{\code{Evaluate}}: runs all benchmark sub-pipelines and scores the performance of each tool
\item \strong{\code{AddLayout}}: allows you to add a separate 2-dimensional layout of the input dataset or to use an existing projection (produced in the evaluation) as a visualisation layout.
}
}
