% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Visual_Heatmaps.R
\name{PlotPopulationHeatmap}
\alias{PlotPopulationHeatmap}
\title{Create a heatmap of expression value medians per population}
\usage{
PlotPopulationHeatmap(benchmark)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}, as generated by the constructor \code{Benchmark} and evaluated using \code{Evaluate}}
}
\description{
Creates a heatmap showing expression value medians per each marker per each labelled population.
}
