% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_WhatIs.R
\name{WhatIs}
\alias{WhatIs}
\title{Inspect a specific cluster or annotated population}
\usage{
WhatIs(
  benchmark,
  idx.subpipeline = NULL,
  idx.n_param = NULL,
  idx.run = 1,
  cluster = NULL,
  population = NULL,
  pointsize = 0.1
)
}
\arguments{
\item{benchmark}{object of type \code{Benchmark}}

\item{idx.subpipeline}{integer: index of subpipeline (if a cluster is to be retrieved). Default value is \code{NULL}}

\item{idx.n_param}{integer: \emph{n}-parameter iteration index (if a cluster is to be retrieved and \emph{n}-parameter was specified in the benchmark set-up). Default value is \code{NULL}}

\item{idx.run}{integer: which clustering run to use if clustering was run repeatedly. Default value is \code{1}}

\item{cluster}{integer or integer vector: index (indices) of cluster of interest (if a cluster is to be retrieved). Default value is \code{NULL}}

\item{population}{string or string vector: name (names) of population of interest (if a population is to be retrieved). Default value is \code{NULL}}

\item{pointsize}{numeric: size of points in expression scatter plot. Default value is \code{0.1}}
}
\description{
For a select group of cells (cluster or annotated population), signal intensiry for each parameter (marker) is plotted, as well as the heatmap of inter-parameter Spearman correlations.
If there are multiple clusters or populations, these plots are generated for all of them.
}
